+++ src/main/java/moze_intel/projecte/gameObjs/container/TransmuteContainer.java
@@ -87,8 +87,8 @@ public class TransmuteContainer extends Container
 			return null;
 		}
 		else if (slotIndex >= 10 && slotIndex <= 25)
-		{	
-			int emc = Utils.getEmcValue(newStack);
+		{
+			double emc = Utils.getEmcValue(newStack);
 			
 			int stackSize = 0;
 			
@@ -103,7 +103,7 @@ public class TransmuteContainer extends Container
 		}
 		else if (slotIndex >= 26)
 		{
-			int emc = Utils.getEmcValue(stack);
+			double emc = Utils.getEmcValue(stack);
 			
 			if (emc == 0)
 			{
