+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUICondenser.java
@@ -34,14 +34,14 @@ public class GUICondenser extends GuiContainer
 		
 		this.drawTexturedModalRect(x, y, 0, 0, xSize, ySize);
 		
-		int progress = tile.getProgressScaled();
-		this.drawTexturedModalRect(x + 33, y + 10, 0, 235, progress, 10);
+		double progress = tile.getProgressScaled();
+		this.drawTexturedModalRect(x + 33, y + 10, 0, 235, (int)progress, 10);
 	}
 	
 	@Override
 	protected void drawGuiContainerForegroundLayer(int var1, int var2) 
 	{
-		int toDisplay = tile.displayEmc > tile.requiredEmc ? tile.requiredEmc : tile.displayEmc;
-		this.fontRendererObj.drawString(Integer.toString(toDisplay), 140, 10, 4210752);
+		double toDisplay = tile.displayEmc > tile.requiredEmc ? tile.requiredEmc : tile.displayEmc;
+		this.fontRendererObj.drawString(Double.toString(toDisplay), 140, 10, 4210752);
 	}
 }
