+++ src/main/java/moze_intel/projecte/network/packets/CollectorSyncPKT.java
@@ -11,15 +11,15 @@ import cpw.mods.fml.common.network.simpleimpl.MessageContext;
 
 public class CollectorSyncPKT implements IMessage, IMessageHandler<CollectorSyncPKT, IMessage>
 {
-	private int displayEmc;
-	private int displayKleinCharge;
+	private double displayEmc;
+	private double displayKleinCharge;
 	private int x;
 	private int y;
 	private int z;
 	
 	public CollectorSyncPKT() {}
 	
-	public CollectorSyncPKT(int displayEmc, int displayKleinCharge, int x, int y, int z) 
+	public CollectorSyncPKT(double displayEmc, double displayKleinCharge, int x, int y, int z)
 	{
 		this.displayEmc = displayEmc;
 		this.displayKleinCharge = displayKleinCharge;
@@ -50,8 +50,8 @@ public class CollectorSyncPKT implements IMessage, IMessageHandler<CollectorSync
 	@Override
 	public void fromBytes(ByteBuf buf) 
 	{
-		displayEmc = buf.readInt();
-		displayKleinCharge = buf.readInt();
+		displayEmc = buf.readDouble();
+		displayKleinCharge = buf.readDouble();
 		x = buf.readInt();
 		y = buf.readInt();
 		z = buf.readInt();
@@ -60,8 +60,8 @@ public class CollectorSyncPKT implements IMessage, IMessageHandler<CollectorSync
 	@Override
 	public void toBytes(ByteBuf buf) 
 	{
-		buf.writeInt(displayEmc);
-		buf.writeInt(displayKleinCharge);
+		buf.writeDouble(displayEmc);
+		buf.writeDouble(displayKleinCharge);
 		buf.writeInt(x);
 		buf.writeInt(y);
 		buf.writeInt(z);
